+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - `Fix` *Config* — User config now has higher priority than internal settings
 - `New` — Ability to work with Block Actions and Inline Toolbar from the keyboard by Tab.
+- `Fix` — Fix placeholder appearance
 
 ### 2.13